requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.13.1
anyio==4.4.0
attrs==23.2.0
bitarray==2.9.2
cbor2==5.6.4
certifi==2024.6.2
charset-normalizer==3.3.2
ckzg==1.0.2
colorama==0.4.6
crytic-compile==0.3.7
cytoolz==0.12.3
distro==1.9.0
et-xmlfile==1.1.0
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.3.1
eth-utils==4.1.1
eth_abi==5.1.0
frozenlist==1.4.1
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Library==0.0.0
lru-dict==1.2.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
numpy==2.0.0
openai==1.34.0
OpenAIAuth==3.0.0
openpyxl==3.1.4
packaging==23.2
pandas==2.2.2
parsimonious==0.10.0
prettytable==3.10.0
prompt_toolkit==3.0.47
protobuf==5.27.1
py-solc-x==2.0.3
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyunormalize==15.1.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
revChatGPT==6.8.6
rich==13.7.1
rlp==4.0.1
rpds-py==0.18.1
scikit-learn==1.5.0
scipy==1.13.1
simplejson==3.19.2
six==1.16.0
slither-analyzer==0.10.3
sniffio==1.3.1
socksio==1.0.0
solc-select==1.0.4
threadpoolctl==3.5.0
tiktoken==0.7.0
tls-client==1.0.1
toolz==0.12.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
web3==6.19.0
websockets==12.0
yarl==1.9.4
